backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.16
mcp>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
websockets>=12.0
aiofiles>=23.0.0
